Broadcast when a processed request is done/failed
So projects can know when cached requests have been processed.

Pass the response if successful, error if failed, and always pass the
original request config object (what was cached).

In the future the http request interceptor could reject a unique id that
the project can then use to track the request when it gets processed.
Kinda similar to how `$on` returns an off function, and `$timeout`
returns a cancel function: we can return something to track the request.